+++ src/Database/Query/MultiInsertQuery.php
@@ -50,7 +50,7 @@ class MultiInsertQuery extends AbstractQuery {
 
 		$stmt->execute($this->getParamtersToBind());
 
-		$this->result = Database::getDbh()->lastInsertId();
+		$this->result = (int)Database::getDbh()->lastInsertId();
 
 		self::$totalQueries++;
